Change design for iPad design
WHAT:
- Update some screens for iPad devices
