+++ Play4Padel/Play4Padel/Common/DataModel/DeviceIdiom.swift
@@ -1,0 +1,16 @@
+//
+//  DeviceIdiom.swift
+//  Play4Padel
+//
+//  Created by Carlos Gonzalez on 12/1/25.
+//
+
+import SwiftUI
+
+@MainActor
+@propertyWrapper
+struct DeviceIdiom {
+    var wrappedValue: UIUserInterfaceIdiom {
+        UIDevice.current.userInterfaceIdiom
+    }
+}
